requirements.txt
django-environ
python-telegram-bot